+++ pages/index.js
@@ -6,14 +6,18 @@ import styles from '../styles/Home.module.css'
 
 const { log, error } = console;
 
+const internals = {
+  reSpace: new RegExp(/\s/g),
+};
+
 export default function Home() {
-  const [searchResults, setSearchResults] = useState([]);
+  const [searchResults, setSearchResults] = useState(null);
   const [isSearching, setIsSearching] = useState(false);
 
   async function handleFormSubmit({ term }) {
-    const q = term.trim();
+    const q = term.replace(internals.reSpace, '');
 
-    setSearchResults([]);
+    setSearchResults(null);
 
     if (!q) return;
 
@@ -22,7 +26,10 @@ export default function Home() {
     try {
       const res = await fetch(`/api/search?q=${q}`);
       const data = await res.json();
-      setSearchResults(data.search.nodes);
+      setSearchResults({
+        search: data.search,
+        term: q
+      });
     } catch (err) {
       error(err);
     } finally {
@@ -47,19 +54,42 @@ export default function Home() {
           loading={isSearching}
           autoFocus={true}
           onFormSubmit={handleFormSubmit}
-          onClearForm={() => setSearchResults([])}
+          onClearForm={() => setSearchResults(null)}
         />
       </div>
 
-      <ul className={styles.searchResultsList}>
-        {searchResults.map((user) => {
-          return (
-            <li key={user.id}>
-              <UserLinkCard data={user}/>
-            </li>
-          );
-        })}
-      </ul>
+      {Boolean(searchResults?.search?.userCount) && (
+        <ul className={styles.searchResultsList}>
+          {searchResults.search.nodes.map((user) => {
+            return (
+              <li key={user.id}>
+                <UserLinkCard data={user}/>
+              </li>
+            );
+          })}
+        </ul>
+      )}
+
+      {Boolean(searchResults?.search?.userCount === 0) && (
+        <div className={styles.noSearchResults}>
+          <p className='mb05'>
+            Your search for “{searchResults.term}” didn’t return any results.
+          </p>
+          <p className='secondary-text'>
+            <i>
+              Check the username typing or <br/>visit
+              {' '}
+              <a
+                href={`https://github.com/${searchResults.term}`}
+                rel='noopener noreferrer'
+                target='_blank'>GitHub/{searchResults.term}
+              </a>
+              {' '}
+              to see if the user exists.
+            </i>
+          </p>
+        </div>
+      )}
     </div>
   )
 }